requirements.txt
transformers==4.28.1
ml_collections
fschat==0.2.20
tenacity==8.2.3
boto3==1.34.23
openai==1.9.0
google-generativeai
anthropic
datasets
accelerate
matplotlib
sentencepiece
/llm-defense/peft
/llm-defense/just_eval